Move reflection functions into seperate file
