+++ mathesar/database/utils.py
@@ -1,6 +1,5 @@
 from django.conf import settings
 
-from mathesar.models import Schema, Database
 from db.engine import get_connection_string
 
 
@@ -22,22 +21,3 @@ def get_database_key(engine):
 
 def get_non_default_database_keys():
     return [key for key in settings.DATABASES if key != 'default']
-
-
-def update_databases():
-    databases = set(settings.DATABASES)
-    # We only want to track non-django dbs
-    databases.remove('default')
-
-    # Update deleted databases
-    for database in Database.objects.all():
-        if database.name in databases:
-            databases.remove(database.name)
-        else:
-            database.deleted = True
-            Schema.objects.filter(database=database).delete()
-            database.save()
-
-    # Create databases that aren't models yet
-    for database in databases:
-        Database.objects.create(name=database)
