+++ mathesar/utils/tables.py
@@ -1,21 +1,6 @@
-from db.tables import get_table_oids_from_schema, infer_table_column_types, create_mathesar_table, get_oid_from_table
+from db.tables import infer_table_column_types, create_mathesar_table, get_oid_from_table
 from mathesar.database.base import create_mathesar_engine
 from mathesar.models import Table
-
-
-def reflect_tables_from_schema(schema):
-    db_table_oids = {
-        table["oid"]
-        for table in get_table_oids_from_schema(schema.oid, schema._sa_engine)
-    }
-    tables = [
-        Table.objects.get_or_create(oid=oid, schema=schema)
-        for oid in db_table_oids
-    ]
-    for table in Table.objects.all().filter(schema=schema):
-        if table.oid not in db_table_oids:
-            table.delete()
-    return tables
 
 
 def get_table_column_types(table):
